# internal/central/pkg/api/admin/private/api/openapi.yaml
openapi: 3.0.0
info:
  description: Red Hat Advanced Cluster Security (RHACS) Service Fleet Manager Admin
    APIs that can be used by RHACS Managed Service Operations Team.
  title: Red Hat Advanced Cluster Security Service Fleet Manager Admin API
  version: 0.0.3
servers:
- description: Main (production) server
  url: https://api.openshift.com
- description: Staging server
  url: https://api.stage.openshift.com
- description: localhost
  url: http://localhost:8000
- description: current domain
  url: /
...
paths:
...
  /api/rhacs/v1/admin/centrals/{id}/billing:
    patch:
      operationId: changeBillingParameters
      parameters:
      - description: The ID of record
        in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CentralBillingChangeRequest'
        description: Billing target parameters
        required: true
      responses:
        "200":
          description: Billing parameters changed
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Auth token is invalid
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: User is not authorised to access the service
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: No Central found with the specified ID or dynamic clients are
            not configured
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected error occurred
      summary: Change central billing parameters
...
components:
...
  schemas:
...
    CentralBillingChangeRequest:
      example:
        cloud_account_id: cloud_account_id
        product: RHACS
        model: standard
        cloud_provider: cloud_provider
      properties:
        model:
          enum:
          - standard
          - marketplace
          - marketplace-aws
          - marketplace-gcp
          - marketplace-rhm
          - marketplace-azure
          type: string
        cloud_account_id:
          type: string
        cloud_provider:
          type: string
        product:
          enum:
          - RHACS
          type: string
      type: object
...
    Error:
      allOf:
      - $ref: '#/components/schemas/ObjectReference'
      - $ref: '#/components/schemas/Error_allOf'
    ObjectReference:
      properties:
        id:
          type: string
        kind:
          type: string
        href:
          type: string
      type: object
...
    Error_allOf:
      properties:
        code:
          type: string
        reason:
          type: string
        operation_id:
          type: string
...
  securitySchemes:
    Bearer:
      bearerFormat: JWT
      scheme: bearer
      type: http